Issue 70

+++ src/paralleltraversal.cpp
@@ -3,10 +3,8 @@
  * @brief File containing functions for index traversal.
  * @parblock
  * SortMeRNA - next-generation reads filter for metatranscriptomic or total RNA
- * @copyright Copyright (C) 2012-2014 Bonsai Bioinformatics Research Group, LIFL and 
- * INRIA Nord-Europe, France
- * OTU-picking extensions developed in the Knight Lab, BioFrontiers Institute,
- * University of Colorado at Boulder, Boulder, CO
+ * @copyright 2013-15 Bonsai Bioinformatics Research Group
+ * 2014-15 Knight Lab, Department of Pediatrics, UCSD, La Jolla
  *
  * SortMeRNA is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -1659,7 +1657,7 @@ paralleltraversal ( char* inputreads,
     vector<uint32_t> skiplengths_v(3,0);
     skiplengths.push_back(skiplengths_v);
   }
-  for (int32_t i = 0; i < myfiles.size()-1; i++) skiplengths.push_back(skiplengths[0]);
+  for (uint32_t i = 0; i < myfiles.size()-1; i++) skiplengths.push_back(skiplengths[0]);
     
   // add header lines to SAM output file
   preprocess_data( myfiles,
@@ -1812,7 +1810,7 @@ paralleltraversal ( char* inputreads,
     TIME(s);
         
     // mmap the reads file into memory
-    char* raw = (char*)mmap ( 0, partial_file_size, PROT_READ, MAP_SHARED, fd, offset_map );
+    char* raw = (char*)mmap(0, partial_file_size, PROT_READ, MAP_SHARED, fd, offset_map);
     if ( raw == MAP_FAILED )
     {
       close(fd);
@@ -1821,13 +1819,14 @@ paralleltraversal ( char* inputreads,
     }
     // pointer to last character in mmap'd region
     char* end_of_mmap = &raw[partial_file_size-1];
+
     
-    int32_t strs = 0;
+    size_t strs = 0;
     
     // the length of the split read in file part i+1 (from beginning of file part)
-    int32_t reads_offset_f = 0;
+    uint32_t reads_offset_f = 0;
     // the length of the split read in file part i (from end of file part)
-    int32_t reads_offset_e = 0;
+    uint32_t reads_offset_e = 0;
         
     {
       // (FASTA) count the number of strings in a file section
@@ -1874,8 +1873,8 @@ paralleltraversal ( char* inputreads,
         else
         {
           // count the number of strings in the file section
-          for ( uint32_t i = reads_offset_f; i < partial_file_size-reads_offset_e-2; i++ ) if ( raw[i] == '>' ) strs++;
-          
+          for ( size_t i = reads_offset_f; i < partial_file_size-reads_offset_e-2; i++ ) if ( raw[i] == '>' ) strs++;
+
           // the paired-read follows the split read at the top of current file section
           if ( offset_pair_from_top )
           {
@@ -1955,7 +1954,7 @@ paralleltraversal ( char* inputreads,
           if ( (strs%4 == 1) && (raw[partial_file_size-1] == '\n') && (raw[partial_file_size-2] == '\n') ) strs--;
           else
           {
-            fprintf(stderr,"   %sERROR%s: Your FASTQ reads file has an uneven number of lines: %u\n","\033[0;31m","\033[0m",strs);
+            fprintf(stderr,"   %sERROR%s: Your FASTQ reads file has an uneven number of lines: %lu\n","\033[0;31m","\033[0m",strs);
             exit(EXIT_FAILURE);
           }
         }
